Speedup doc generation when running bun run dev
